% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_workday_report.R
\name{get_workday_report}
\alias{get_workday_report}
\title{Retrieve a report via the Workday RaaS}
\usage{
get_workday_report(report_name, organization, username, password,
  params = list(format = "csv"), file_path = tempfile(),
  overwrite = TRUE, url_params = NULL, ...)
}
\arguments{
\item{report_name}{character. The name of the report. It should have the format REPORT_OWNER/REPORT_NAME you will need to pass the full string in.}

\item{organization}{character. The Workday organization. To find yours go to your Workday home page, the organization name is in the ulr \code{https://wd5.myworkday.com/ORGANIZATION_NAME/d/home.htmld}}

\item{username}{character. The username of the user who has access to the report.}

\item{password}{character. The password of the user who has access to the report.}

\item{params}{list. URL parameters to send into \code{httr::GET()}. Defaults to \code{list(format='csv')}. Common params include \code{list(Effective_as_of_date='2018-10-01-07:00')}.}

\item{file_path}{where should the report be saved? Defaults to \code{tempfile()}.}

\item{overwrite}{boolean. Should the be overwritten? Defaults to \code{TRUE}.}

\item{url_params}{Deprecated use params.}

\item{...}{parameters to pass into \code{httr::GET()}.}
}
\description{
Query's the Workday Report As a Service API. A tutorial on how to use this API is located \href{https://docs.workato.com/connectors/workday/workday_raas.html}{here}
This function specifically uses \code{httr::GET(httr::write_disk())} to return the result as a file_path to the saved report.
After you get the response you can use a \code{readr} function to read the result such as \code{readr::read_csv}.
Note that this endpoint has been producing SSL errors. The \code{httr::GET()} request is wrapped in a \code{purrr::safely()} to write the file to disk regardless of error.
}
