% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_workday_report.R
\name{get_workday_report}
\alias{get_workday_report}
\title{Retrieve a report via the Workday RaaS}
\usage{
get_workday_report(report_name, organization, username, password,
  url_params = c(format = "csv"), ...)
}
\arguments{
\item{report_name}{character. The name of the report. It should have the format REPORT_OWNER/REPORT_NAME you will need to pass the full string in.}

\item{organization}{character. The Workday organization. To find yours go to your Workday home page, the organization name is in the ulr \code{https://wd5.myworkday.com/ORGANIZATION_NAME/d/home.htmld}}

\item{username}{character. The username of the user who has access to the report.}

\item{password}{character. The password of the user who has access to the report.}

\item{url_params}{named vector of key value pairs. Defaults to \code{c('format'='csv')}. Common params include \code{c('Effective_as_of_date'='2018-10-01-07:00')}.}

\item{...}{parameters to pass into \code{RCurl::getURL()}}
}
\description{
Query's the Workday Report As a Service API. A tutorial on how to use this API is located \href{https://docs.workato.com/connectors/workday/workday_raas.html}{here}
This function specifically uses \code{RCurl::getURL()} to return the result as text or binary. After you get the response you can use a \code{readr} function to read the result such as
\code{readr::read_csv}.
}
